refactor(npm-scripts): move createBridges to utils
It's not an script.
